+++ CHANGELOG.md
@@ -4,6 +4,5 @@
 * To use the mod as before with the Global Datapacks option enabled, you need to move the ```datapacks``` path from optional datapacks to required datapacks
 
 ### New changes
-* The mod is working on the server again
-* Added the ability to make features global
-* Translation corrected
+* Fixed a bug where required datapacks were incorrectly displayed in the selection menu during world creation
+* Datapacks are now copied to the correct directory, and not only optional datapacks but also required ones are copied